OPT: remove index.js
