+++ components/check/index.js
@@ -1,23 +1,33 @@
 import React, { PureComponent } from 'react'
 import PropTypes from 'prop-types'
 
-import { trimList } from '../util'
-import { getOptionLabel, getOptionValue, checkOptionByValue } from './util'
-
+import isArray from 'lodash/isArray'
+import isSet from 'lodash/isSet'
 import isEqual from 'lodash/isEqual'
 
+import Icon from '../icon'
+
+import {
+  getOptionLabel,
+  getOptionValue,
+  checkOptionByValueList,
+  convertValueListToSet,
+  trimList,
+} from '../util'
+
+import './index.styl'
 /**
- * <Radio>
+ * <Check>
  */
-export class Radio extends PureComponent {
+export default class Check extends PureComponent {
+  state = {
+    prevProps: this.props,
+    isChecked: this.props.isChecked,
+  }
+
   static propTypes = {
     size: PropTypes.oneOf(['regular', 'small']),
     theme: PropTypes.oneOf(['core', 'plain']),
-    className: PropTypes.string,
-
-    label: PropTypes.any,
-    name: PropTypes.string,
-    value: PropTypes.any,
 
     isChecked: PropTypes.bool,
     isDisabled: PropTypes.bool,
@@ -26,23 +36,22 @@ export class Radio extends PureComponent {
 
     onChange: PropTypes.func.isRequired,
     onToggle: PropTypes.func.isRequired,
+
+    label: PropTypes.any,
+    name: PropTypes.string,
+    value: PropTypes.any,
+    className: PropTypes.string,
   }
 
   static defaultProps = {
     size: 'regular',
     theme: 'plain',
-
     isChecked: false,
     label: '',
     className: '',
 
     onChange: () => null,
     onToggle: () => null,
-  }
-
-  state = {
-    prevProps: this.props,
-    isChecked: this.props.isChecked,
   }
 
   static getDerivedStateFromProps(props, { prevProps, isChecked }) {
@@ -67,16 +76,15 @@ export class Radio extends PureComponent {
   }
 
   onToggle = () => {
-    const { name, value, label, onToggle, onChange } = this.props
+    const { name, value, label, onChange, onToggle } = this.props
     const { isChecked } = this.state
     const { canToggle } = this
 
-    const result = canToggle ? true : isChecked
+    const willBeChecked =  canToggle ? !isChecked : isChecked
 
-    this.setState({ isChecked: result })
-
-    onToggle(result, name, value || label)
-    onChange(name, value || label, result)
+    this.setState({ isChecked: willBeChecked })
+    onToggle(willBeChecked, name, value || label)
+    onChange(name, value || label, willBeChecked)
   }
 
   render() {
@@ -86,26 +94,23 @@ export class Radio extends PureComponent {
 
     return (
       <label
-        className={
-          trimList([
-            theme === 'core' ? 'CoreRadio' : 'Radio',
-            size,
-            className,
-            isChecked && 'is-checked',
-            isDisabled && 'is-disabled',
-            readOnly && 'readonly',
-          ])
-        }
+        className={trimList([
+          theme === 'core' ? 'CoreCheck' : 'Check',
+          size,
+          className,
+          isChecked && 'is-checked',
+          isDisabled && 'is-disabled',
+          readOnly && 'readonly',
+        ])}
       >
         <input
-          type="radio"
+          type="checkbox"
           defaultChecked={isChecked}
           disabled={isDisabled}
           name={name}
-          onClick={this.onToggle}
+          onChange={this.onToggle}
         />
-
-        <span className="Check-state" />
+        <span className="Check-state"><Icon type="dora" name="check" /></span>
         <span className="Check-label">{ label }</span>
       </label>
     )
@@ -113,22 +118,24 @@ export class Radio extends PureComponent {
 }
 
 /**
- * <RadioGroup>
+ * <CheckGroup>
  */
-export class RadioGroup extends PureComponent {
+export class CheckGroup extends PureComponent {
   name = this.props.name || Math.random().toString(36).substring(2, 15)
 
   state = {
     prevProps: this.props,
-    value: this.props.value,
+    valueList: convertValueListToSet(this.props.valueList),
   }
 
   static propTypes = {
+    name: PropTypes.string,
     size: PropTypes.oneOf(['regular', 'small']),
     theme: PropTypes.oneOf(['core', 'plain']),
     className: PropTypes.string,
 
-    name: PropTypes.string,
+    onChange: PropTypes.func.isRequired,
+    onToggle: PropTypes.func.isRequired,
 
     optionList: PropTypes.arrayOf(
       PropTypes.oneOfType([
@@ -142,33 +149,29 @@ export class RadioGroup extends PureComponent {
       ])
     ).isRequired,
 
-    value: PropTypes.oneOfType([
-      PropTypes.number,
-      PropTypes.string,
+    valueList: PropTypes.oneOfType([
+      PropTypes.instanceOf(Set),
+      PropTypes.array,
     ]),
 
     isDisabled: PropTypes.bool,
     disabled: PropTypes.bool,
     readOnly: PropTypes.bool,
-
-    onChange: PropTypes.func.isRequired,
-    onToggle: PropTypes.func.isRequired,
   }
 
   static defaultProps = {
     size: 'regular',
     theme: 'plain',
-
     className: '',
     optionList: [],
-
     onChange: () => null,
     onToggle: () => null,
+    isDisabled: false,
   }
 
-  static getDerivedStateFromProps(props, { prevProps, value }) {
+  static getDerivedStateFromProps(props, { prevProps, valueList }) {
     if (!isEqual(prevProps, props)) {
-      return { prevProps: props, value: props.value }
+      return { prevProps: props, valueList: props.valueList }
     }
     return null
   }
@@ -187,25 +190,42 @@ export class RadioGroup extends PureComponent {
     return !isDisabled && !readOnly
   }
 
-  createOnChangeHandler = (name, value, idx) => () => {
-    const { onToggle, onChange } = this.props
-    const { value: originalValue } = this.state
+  createOnChangeHandler = (name, opt) => () => {
+    const { optionList, onToggle, onChange } = this.props
+    const { valueList } = this.state
     const { canToggle } = this
 
-    const result = canToggle ? value : originalValue
 
-    this.setState({ value: result })
-    onToggle(result, name)
-    onChange({ name, value: result, idx })
+    if (!canToggle) {
+      const currentValueList = Array.from(valueList)
+      const currentIdxList = currentValueList.map(v => optionList.findIndex(o => getOptionValue(o) === v))
+      onToggle(currentValueList, name)
+      onChange({ name, valueList: currentValueList, idxList: currentIdxList  })
+      return
+    }
+
+    const resultValueList = new Set(valueList)
+    const optionValue = getOptionValue(opt)
+
+    const toggleAction = resultValueList.has(optionValue) ? 'delete' : 'add'
+    resultValueList[toggleAction](optionValue)
+
+    const nextValueList = Array.from(resultValueList)
+    const nextIdxList = nextValueList.map(v => optionList.findIndex(o => getOptionValue(o) === v))
+
+    this.setState({ valueList: resultValueList })
+
+    onToggle(nextValueList, name)
+    onChange({ name, valueList: nextValueList, idxList: nextIdxList })
   }
 
   render() {
     const { size, theme, className, optionList } = this.props
-    const { value } = this.state
+    const { valueList } = this.state
     const { name, isDisabled, readOnly } = this
 
     const klass = trimList([
-      theme === 'core' ? 'CoreRadioGroup' : 'RadioGroup',
+      theme === 'core' ? 'CoreCheckGroup' : 'CheckGroup',
       size,
       className,
       isDisabled && 'is-disabled',
@@ -215,23 +235,22 @@ export class RadioGroup extends PureComponent {
     return (
       <span className={klass}>
       {
-        optionList
-        .map((opt, idx) => opt && (
-          <Radio
+        optionList.map((opt, idx) => opt && (
+          <Check
             key={idx}
             name={name}
+            label={getOptionLabel(opt)}
+
             size={size}
             theme={theme}
 
-            label={getOptionLabel(opt)}
-            type="radio"
-            isChecked={checkOptionByValue(opt, value)}
             isDisabled={isDisabled || opt.isDisabled}
             readOnly={readOnly}
+            isChecked={checkOptionByValueList(opt, valueList)}
 
             onChange={
               !(isDisabled || opt.isDisabled)
-              ? this.createOnChangeHandler(name, getOptionValue(opt), idx)
+              ? this.createOnChangeHandler(name, opt)
               : undefined
             }
           />
@@ -242,10 +261,10 @@ export class RadioGroup extends PureComponent {
   }
 }
 
-export function CoreRadio(props) {
-  return <Radio {...props} theme="core" />
+export function CoreCheck(props) {
+  return <Check {...props} theme="core" />
 }
 
-export function CoreRadioGroup(props) {
-  return <RadioGroup {...props} theme="core" />
+export function CoreCheckGroup(props) {
+  return <CheckGroup {...props} theme="core" />
 }
