+++ components/confirmInputNumber/index.styl
@@ -1,3 +1,5 @@
+@import '../util/form'
+
 .InputNumber
 .CoreInputNumber
   &.is-hover
