+++ components/core-modal/index.styl
@@ -1,4 +1,4 @@
-@import '../util/index'
+@import '../util/form'
 
 .CoreModalPortal
   &:empty
