+++ components/ellipsis/ellipsis-extends.styl
