+++ components/ellipsis/index.styl
