+++ components/emailInput/index.styl
@@ -1,3 +1,5 @@
+@import '../util/form'
+
 .InputEmail
 .CoreInputEmail
   &:not(.unstyled).isnt-valid input
