--- components/form/index.js
@@ -1,19 +1,0 @@
-import './index.styl'
-
-export { FormEntry, FormLabel } from './FormEntry'
-
-export {
-  Input, PanelInput, Textarea, PanelTextarea,
-  CoreInput, CoreTextarea,
-
-  InputEmail, PanelInputEmail, CoreInputEmail,
-
-  InputNumber, SelectNumber, CoreInputNumber, CoreSelectNumber,
-
-  ConfirmInputNumber, CoreConfirmInputNumber,
-  PanelInputNumber, PanelSelectNumber,
-} from './Input'
-
-export { Check, CheckGroup, CoreCheck, CoreCheckGroup } from './Check'
-export { Radio, RadioGroup, CoreRadio, CoreRadioGroup } from './Radio'
-export { Select, PanelSelect, CoreSelect } from './Select'
