+++ components/formEntry/index.js
@@ -2,10 +2,12 @@ import React, { Fragment } from 'react'
 import PropTypes from 'prop-types'
 
 import { trimList } from '../util'
+
+import './index.styl'
 /**
  * <FormEntry>
  */
-export function FormEntry({
+export default function FormEntry({
   className,
   name: key, children: val,
   type,
