+++ components/formEntry/index.styl
@@ -1,3 +1,5 @@
+@import '../util/form'
+
 .FormEntry
   &
     display flex
