--- components/index.js
@@ -1,14 +1,0 @@
-export { default as Root } from './root'
-
-export * from './button'
-export { default as Dropdown } from './dropdown'
-export * from './form'
-export { default as Guide } from './guide'
-export { default as Icon } from './icon'
-export { default as Modal } from './modal'
-export { default as CoreModal } from './core-modal'
-export { default as Overlay } from './overlay'
-export { default as SVG } from './svg'
-export { default as Switch } from './switch'
-export * from './text'
-export * from './util'
