+++ components/input/index.styl
@@ -1,3 +1,5 @@
+@import '../util/form'
+
 $Input
   &
     display inline-block
