+++ components/numberInput/index.styl
@@ -1,3 +1,5 @@
+@import '../util/form'
+
 .InputNumber
 .CoreInputNumber
   $padding-left = .5em
@@ -61,7 +63,7 @@
       /.InputNumber.is-menu-open .caret svg
         transform rotate(180deg)
 
-     button
+    button
       &
         display block
         margin 0
