+++ components/radio/index.styl
@@ -1,7 +1,7 @@
+@import '../check/index'
+
 .Radio
 .CoreRadio
-  @extends .Check
-
   .Check-state
     padding 1px
     border-radius 50%
@@ -20,9 +20,3 @@
       background-color #fff
       border-radius 50%
 
-.CoreRadio
-  @extends .CoreCheck
-
-.RadioGroup
-.CoreRadioGroup
-  @extends .CheckGroup
