+++ components/select/index.styl
@@ -1,3 +1,5 @@
+@import '../util/form'
+
 .Select
 .CoreSelect
   $padding-left = .5em
