--- components/text/index.js
@@ -1,5 +1,0 @@
-import './index.styl'
-
-export * from './Ellipsis'
-export Tooltip from './Tooltip'
-export { CoreTooltip } from './Tooltip'
