--- components/text/index.styl
@@ -1,4 +1,0 @@
-@import '../util/index'
-
-@import './Tooltip'
-@import './Ellipsis'
