+++ components/tooltip/index.styl
@@ -1,3 +1,5 @@
+@import '../util/index'
+
 div.Tooltip
   width -moz-fit-content
   width -webkit-fit-content
