+++ components/util/form.styl
@@ -1,4 +1,4 @@
-@import '../util/index'
+@import './index.styl'
 
 $FormEntry-key-width = 7em
 $FormEntry-min-height = 34px
@@ -52,14 +52,3 @@ $cl-InputNumber-desc = $cl-tertiary-grey
 $cl-Check-text = $cl-secondary-grey
 $cl-Select-text = $cl-secondary-grey
 
-@import './FormEntry'
-
-@import './Input'
-@import './InputEmail'
-@import './InputNumber'
-@import './ConfirmInputNumber'
-
-@import './Check'
-@import './Radio'
-
-@import './Select'
