+++ components/util/index.js
@@ -1,12 +1,14 @@
 import { compact, isBoolean } from 'lodash'
-
+import isNumber from 'lodash/isNumber'
+import isString from 'lodash/isString'
+import { isValidElement } from 'react'
 import './polyfill'
 
 import './index.styl'
 
 export * as SVG from './svg'
 
-export const EMAIL_REGEX = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/
+export const EMAIL_REGEX = /^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z0-9-]{2,63}$/i
 
 const IBOT = { OPEN_MODAL_STACK: [] }
 
@@ -124,4 +126,40 @@ export function toggleGlobalScroll(expected) {
   }
 
   return is
+}
+
+export function getOptionLabel(it) {
+  return (
+    isString(it) || isNumber(it) || isValidElement(it)
+    ? it
+    : it.label || it.value
+    ? it.label || it.value
+    : undefined
+  )
+}
+
+export function getOptionValue(it) {
+  return (
+    isString(it) || isNumber(it)
+    ? String(it)
+    : it.value || it.label
+    ? String(it.value || it.label)
+    : undefined
+  )
+}
+
+export function convertValueListToSet(valueList) {
+  return new Set(Array.from(valueList || []).map(String))
+}
+
+export function checkOptionByValue(it, value) {
+  return !!value && getOptionValue(it) === String(value)
+}
+
+export function checkOptionByValueList(it, valueList) {
+  return convertValueListToSet(valueList).has(getOptionValue(it))
+}
+
+export function setNumberValue(value) {
+  return value !== '' && isNumber(Number(value)) ? Number(value) : ''
 }
