+++ stories/form/CoreFormComponent.js
@@ -1,19 +1,16 @@
 import React, { PureComponent } from 'react'
 import { action } from '@storybook/react'
 
+import Root from '../../components/root'
 import {
-  Root,
-
-  CoreCheck, CoreCheckGroup,
-  CoreRadio, CoreRadioGroup,
-
   CoreInput,
   CoreInputEmail,
   CoreInputNumber,
   CoreConfirmInputNumber,
-
-  CoreSelect,
-} from '../../components'
+} from '../../components/input'
+import { CoreRadio, CoreRadioGroup } from '../../components/radio'
+import { CoreSelect } from '../../components/select'
+import { CoreCheck, CoreCheckGroup } from '../../components/check'
 
 export default class CoreFormComponent extends PureComponent {
   state = {
@@ -44,7 +41,7 @@ export default class CoreFormComponent extends PureComponent {
         <div className="single">
           <CoreCheck name="天干" label="甲" />
           <CoreCheck name="天干" label="乙" />
-          <CoreCheck name="天干" label="丙" /> 
+          <CoreCheck name="天干" label="丙" />
         </div>
 
         <div className="single">
@@ -53,7 +50,7 @@ export default class CoreFormComponent extends PureComponent {
           <CoreRadio name="天干" label="丙" isChecked={tiangan === '丙'} onToggle={this.onToggleTiangan} />
         </div>
 
-        <CoreCheckGroup name="latin" optionList={['A', 'B', 'C']} /> 
+        <CoreCheckGroup name="latin" optionList={['A', 'B', 'C']} />
         <CoreRadioGroup name="latin" optionList={['A', 'B', 'C']} />
 
         <div className="row">
