+++ stories/form/Select.js
@@ -1,7 +1,10 @@
 import React from 'react'
 import { action } from '@storybook/addon-actions'
 
-import { Root, Icon, Button, Select } from '../../components'
+import Root from '../../components/root'
+import Button from '../../components/button'
+import Select from '../../components/select'
+import Icon from '../../components/icon'
 
 const MENU_THEME_LIST = ['plain', 'core', 'check']
