+++ stories/icon.js
@@ -2,7 +2,9 @@ import React, { PureComponent } from 'react'
 import { storiesOf } from '@storybook/react'
 import shuffle from 'lodash/shuffle'
 
-import { Root, Icon, Button } from '../components'
+import Root from '../components/root'
+import Button from '../components/button'
+import Icon from '../components/icon'
 
 const { DuoIcon } = Icon
