+++ stories/switch.js
@@ -2,7 +2,8 @@ import React from 'react'
 import { storiesOf } from '@storybook/react'
 import { action } from '@storybook/addon-actions'
 
-import { Root, Switch } from '../components'
+import Root from '../components/root'
+import Switch from '../components/switch'
 
 storiesOf('Switch', module)
 .add('On/off', () => (
